improve editor ux for memberplan editing. show only periodicities actually activated in the settings.
