+++ apps/editor/src/app/locales/de.json
@@ -1684,8 +1684,9 @@
       "migratePMHelptext": "Migriere die Abos dieses Planes zu einer neuen Zahlungsmethode. Abos, die noch automatisch abgebucht werden, werden nicht migriert. Die Migration findet erst dann statt, wenn ein User selbst tätig wird und in seinem Profil etwa eine offene Rechnung bezahlen will oder ein Abo frühzeitig verlängert.",
       "amountPerMonthTarget": "Ziel-Betrag / Monat",
       "amountPerMonthTargetHelpText": "Der monatliche Zielbetrag legt den gewünschten Abopreis fest, den die Benutzer:innen bezahlen sollten. Er muss höher als der Minimalbetrag sein und definiert typischerweise die Voreinstellung im Zahlungsformular.",
-      "periodicityPricing": "Preise pro Zahlungsintervall",
+      "periodicityPricing": "Preise pro aktiviertes Zahlungsintervall",
       "periodicityPricingHelpText": "Setze exakte Preise (min / Ziel / max) pro Zahlungsintervall. Ohne eigene Preise wird der Betrag aus dem Monatspreis hochgerechnet. Der Rabatt gegenüber dem hochgerechneten Monatspreis wird automatisch berechnet.",
+      "periodicityPricingHelpTextNoMonthly": "Setze exakte Preise (min / Ziel / max) pro Zahlungsintervall. Da kein monatliches Zahlungsintervall aktiviert ist, wird der interne monatliche Referenzpreis automatisch aus dem günstigsten Intervall berechnet.",
       "periodicityPricingNoPeriodicities": "Aktiviere zuerst Zahlungsintervalle in den Zahlungseinstellungen, um intervall-spezifische Preise zu konfigurieren.",
       "periodicityPricingDelta": "Rabatt gegenüber Monatspreis: {{delta}}",
       "periodicityPricingCustom": "Eigene Preise",
@@ -1695,6 +1696,7 @@
       "periodicityPricingMax": "Maximalbetrag pro Periode",
       "amountPerMonthMinHelpText": "Der minimale monatliche Betrag ist der eigentliche Abopreis, der mindestens bezahlt werden muss. Die Benutzer:innen können indes beim Aboabschluss freiwillig mehr bezahlen.",
       "amountPerMonthMaxHelpText": "Der maximale monatliche Betrag ist der Höchstbetrag, den die Benutzer:innen beim Aboabschluss bezahlen können.",
+      "redirectionsTitle": "Weiterleitungen",
       "confirmationPage": "Weiterleitung zur Seite bei bestehendem User (Retargeting)",
       "confirmationPageHelptext": "Wenn beim Lösen eines Abos die User-ID bekannt ist, erhält die Person eine E-Mail, um das Abo zu bezahlen. Darauf könnte auf dieser Seite hingewiesen werden.",
       "productType": "Produkttyp",
