+++ apps/editor/src/app/locales/en.json
@@ -1742,8 +1742,9 @@
       "migratePMHelptext": "Migrate the subscriptions of this plan to a new Payment Method. Subscriptions that are still debited automatically will not be migrated. The migration only takes place when a user takes action themselves and wants to pay an open invoice in their profile or renew a subscription early.",
       "amountPerMonthTarget": "Target amount / month",
       "amountPerMonthTargetHelpText": "The monthly target amount defines the desired memberplan price that users should pay. It must be higher than the minimum amount and typically defines the default setting in the payment form.",
-      "periodicityPricing": "Prices per payment interval",
+      "periodicityPricing": "Prices per enabled payment interval",
       "periodicityPricingHelpText": "Set exact prices (min / target / max) per payment interval. Without custom prices the amount is derived from the monthly price. The discount compared to the derived monthly price is calculated automatically.",
+      "periodicityPricingHelpTextNoMonthly": "Set exact prices (min / target / max) per payment interval. Since no monthly payment interval is enabled, the internal monthly reference price is calculated automatically from the cheapest interval.",
       "periodicityPricingNoPeriodicities": "Enable payment intervals in the payment settings first to configure interval-specific prices.",
       "periodicityPricingDelta": "Discount compared to monthly price: {{delta}}",
       "periodicityPricingCustom": "Custom prices",
@@ -1753,6 +1754,7 @@
       "periodicityPricingMax": "Maximum amount per period",
       "amountPerMonthMinHelpText": "The minimum monthly amount is the actual memberplan price that must be paid as a minimum. However, users can voluntarily pay more when subscribing.",
       "amountPerMonthMaxHelpText": "The maximum monthly amount is the highest amount users can pay when subscribing.",
+      "redirectionsTitle": "Redirects",
       "confirmationPage": "Redirect to page for existing users (retargeting)",
       "confirmationPageHelptext": "If the user ID is known when a subscription is ordered, the person will receive an e-mail to pay for the subscription. This could be indicated on this page.",
       "productType": "Product type",
