+++ apps/editor/src/app/locales/fr.json
@@ -1576,6 +1576,8 @@
       "editTitle": "Modifier l'abonnement",
       "memberPlanMonthlyAmount": "Montant mensuel minimum ({{currency}})",
       "monthlyAmount": "Montant mensuel",
+      "monthlyAmountEquivalent": "Correspond à {{currency}} {{amount}} par mois",
+      "periodAmount": "Montant par intervalle de paiement",
       "paidUntil": "Payé jusqu'à",
       "paymentMethod": "Mode de paiement",
       "renewNow": "Prolonger maintenant",
@@ -1632,13 +1634,27 @@
       "trialSubscriptionNotPossible": "Le plan d'abonnement doit rester extensible, car un renouvellement automatique est paramétré dans un paramètre de paiement. Tu peux bien sûr changer cela.",
       "migratePMTitle": "Migrer les abonnements vers le mode de paiement cible",
       "migratePMHelptext": "Migre les abonnements de ce plan vers un nouveau mode de paiement. Les abonnements qui sont encore débités automatiquement ne sont pas migrés. La migration n'a lieu que lorsqu'un utilisateur agit lui-même et veut par exemple payer une facture ouverte dans son profil ou prolonger un abonnement de manière anticipée.",
+      "periodicityPricing": "Prix par intervalle de paiement activé",
+      "periodicityPricingHelpText": "Définis des prix exacts (min / cible / max) par intervalle de paiement. Sans prix personnalisés, le montant est extrapolé à partir du prix mensuel. Le rabais par rapport au prix mensuel extrapolé est calculé automatiquement.",
+      "periodicityPricingHelpTextNoMonthly": "Définis des prix exacts (min / cible / max) par intervalle de paiement. Comme aucun intervalle de paiement mensuel n'est activé, le prix mensuel de référence interne est calculé automatiquement à partir de l'intervalle le plus avantageux.",
+      "periodicityPricingNoPeriodicities": "Active d'abord des intervalles de paiement dans les paramètres de paiement pour configurer des prix spécifiques aux intervalles.",
+      "periodicityPricingDelta": "Rabais par rapport au prix mensuel : {{delta}}",
+      "periodicityPricingCustom": "Prix personnalisés",
+      "periodicityPricingDerived": "À partir du prix mensuel",
+      "periodicityPricingMin": "Montant minimal par période",
+      "periodicityPricingTarget": "Montant cible par période",
+      "periodicityPricingMax": "Montant maximal par période",
+      "redirectionsTitle": "Redirections",
       "confirmationPage": "Redirection vers la page en cas d'utilisateur existant (retargeting)",
       "confirmationPageHelptext": "Si l'identifiant de l'utilisateur est connu lors de la souscription d'un abonnement, la personne reçoit un e-mail pour payer l'abonnement. Cela pourrait être signalé sur cette page.",
       "productType": "Type de produit",
       "productTypeHelpText": "Choisissez si le plan doit être proposé comme abonnement ou don. ",
       "productTypeSubscription": "Abonnement",
       "productTypeDonation": "Don",
-      "amountPerMonthMaxHelpText": "Le montant maximum par mois qu'un utilisateur peut payer pour ce plan d'abonnement."
+      "amountPerMonthMaxHelpText": "Le montant maximum par mois qu'un utilisateur peut payer pour ce plan d'abonnement.",
+      "amountPerMonthTarget": "Montant cible / mois",
+      "amountPerMonthMinHelpText": "Le montant mensuel minimum est le prix effectif de l'abonnement qui doit être payé au minimum. Les utilisateur·trice·s peuvent toutefois payer volontairement davantage lors de la souscription.",
+      "amountPerMonthTargetHelpText": "Le montant mensuel cible définit le prix d'abonnement souhaité que les utilisateur·trice·s devraient payer. Il doit être supérieur au montant minimum et définit typiquement la valeur par défaut dans le formulaire de paiement."
     },
     "subscriptionEditView": {
       "additionalSettingsTitle": "Paramètres avancés",
