+++ libs/membership/editor/src/lib/memberplan/memberplan-form.tsx
@@ -1,10 +1,10 @@
 import React, { Dispatch, SetStateAction, useMemo, useState } from 'react';
 import {
+  Currency,
   FullMemberPlanFragment,
   FullPaymentMethodFragment,
   FullImageFragment,
   PaymentMethod,
-  Currency,
   ProductType,
   FullAvailablePaymentMethodFragment,
 } from '@wepublish/editor/api';
@@ -30,7 +30,6 @@ import { slugify } from '@wepublish/utils';
 import {
   ALL_PAYMENT_PERIODICITIES,
   ChooseEditImage,
-  CurrencyInput,
   ImageEditPanel,
   ImageSelectPanel,
   ListInput,
@@ -40,7 +39,7 @@ import {
   SelectPage,
 } from '@wepublish/ui/editor';
 import { MdAutoFixHigh, MdCheck } from 'react-icons/md';
-import { MemberPlanPeriodicityPricing } from './memberplan-periodicity-pricing';
+import { MemberPlanPricing } from './memberplan-pricing';
 import { Alert } from '@mui/material';
 import styled from '@emotion/styled';
 
@@ -63,6 +62,21 @@ const PanelWidth100 = styled(Panel)`
 
 const RowPaddingTop = styled(Row)`
   padding-top: 12px;
+`;
+
+const DividerTextAlignLeft = styled(Divider)`
+  &&& {
+    margin-left: -5px;
+    margin-right: -5px;
+  }
+
+  &&&::before {
+    content: none;
+  }
+
+  &&& > .rs-divider-inner-text {
+    padding-left: 0;
+  }
 `;
 
 interface MemberPlanFormProps {
@@ -369,9 +383,12 @@ export function MemberPlanForm({
       </Col>
 
       <Col xs={12}>
-        <Panel bordered>
+        <Panel
+          header={t('memberplanForm.trialSubscription')}
+          bordered
+        >
+          {/* tags */}
           <Row>
-            {/* tags */}
             <Col xs={24}>
               <Form.ControlLabel>{t('memberPlanEdit.tags')}</Form.ControlLabel>
               <TagPicker
@@ -393,9 +410,161 @@ export function MemberPlanForm({
                 }}
               />
             </Col>
+          </Row>
 
-            {/* Currency */}
+          {/* automatically configure trial subscription */}
+          <RowPaddingTop>
             <Col xs={24}>
+              {isTrialSubscription ?
+                <Alert
+                  icon={<MdCheck />}
+                  severity="success"
+                >
+                  {t('memberplanForm.trialMemberplanAlert')}
+                </Alert>
+              : <Button
+                  startIcon={<MdAutoFixHigh />}
+                  onClick={() =>
+                    setExtendable(
+                      false,
+                      memberPlan ? { ...memberPlan, maxCount: 1 } : undefined
+                    )
+                  }
+                  disabled={isTrialSubscription}
+                  color={'green'}
+                >
+                  {t('memberplanForm.configureTrialBtn')}
+                </Button>
+              }
+            </Col>
+          </RowPaddingTop>
+          <RowPaddingTop>
+            {/* extendable */}
+            <Col xs={12}>
+              <Toggle
+                checked={memberPlan?.extendable}
+                onChange={extendable => setExtendable(extendable)}
+              />
+              <FormControlLabelMarginLeft>
+                {t('memberplanForm.extendableToggle')}
+              </FormControlLabelMarginLeft>
+              <HelpText>{t('memberplanForm.extendableHelpText')}</HelpText>
+            </Col>
+            {/* max count */}
+            <Col xs={12}>
+              <ControlLabel>{maxCountLabel}</ControlLabel>
+              <Input
+                placeholder={maxCountLabel}
+                type={'number'}
+                min={0}
+                value={memberPlan?.maxCount || undefined}
+                onChange={maxCount => {
+                  if (!memberPlan) {
+                    return;
+                  }
+                  setMemberPlan({
+                    ...memberPlan,
+                    maxCount: Number(maxCount) || null,
+                  });
+                }}
+              />
+              <HelpText>{maxCountHelpText}</HelpText>
+            </Col>
+          </RowPaddingTop>
+          <RowPaddingTop>
+            <Col xs={12}>
+              <ControlLabel>{t('memberplanForm.migratePMTitle')}</ControlLabel>
+              <Control
+                name="migrateToTargetPaymentMethodID"
+                block
+                virtualized
+                disabled={loading}
+                data={paymentMethods.map(pm => ({
+                  value: pm.id,
+                  label: pm.name,
+                }))}
+                value={memberPlan?.migrateToTargetPaymentMethodID}
+                accepter={SelectPicker}
+                placement="auto"
+                onChange={migrateToTargetPaymentMethodID =>
+                  setMemberPlan({
+                    ...(memberPlan as FullMemberPlanFragment),
+                    migrateToTargetPaymentMethodID:
+                      migrateToTargetPaymentMethodID || null,
+                  })
+                }
+              />
+              <HelpText>{t('memberplanForm.migratePMHelptext')}</HelpText>
+            </Col>
+          </RowPaddingTop>
+
+          {/* redirections */}
+          <DividerTextAlignLeft>
+            {t('memberplanForm.redirectionsTitle')}
+          </DividerTextAlignLeft>
+          <Row>
+            <Form.ControlLabel>
+              {t('memberPlanEdit.successPage')}
+            </Form.ControlLabel>
+            <SelectPage
+              setSelectedPage={successPageId => {
+                if (!memberPlan) {
+                  return;
+                }
+
+                setMemberPlan({ ...memberPlan, successPageId });
+              }}
+              selectedPage={memberPlan?.successPageId}
+              name="successPageId"
+            />
+          </Row>
+
+          <RowPaddingTop>
+            <Form.ControlLabel>
+              {t('memberPlanEdit.failPage')}
+            </Form.ControlLabel>
+            <SelectPage
+              setSelectedPage={failPageId => {
+                if (!memberPlan) {
+                  return;
+                }
+
+                setMemberPlan({ ...memberPlan, failPageId });
+              }}
+              selectedPage={memberPlan?.failPageId}
+              name="failPageId"
+            />
+          </RowPaddingTop>
+
+          <RowPaddingTop>
+            <Form.ControlLabel>
+              {t('memberplanForm.confirmationPage')}
+            </Form.ControlLabel>
+            <SelectPage
+              setSelectedPage={confirmationPageId => {
+                if (!memberPlan) {
+                  return;
+                }
+
+                setMemberPlan({ ...memberPlan, confirmationPageId });
+              }}
+              selectedPage={memberPlan?.confirmationPageId}
+              name="failPageId"
+            />
+          </RowPaddingTop>
+          <HelpText>{t('memberplanForm.confirmationPageHelptext')}</HelpText>
+        </Panel>
+      </Col>
+
+      {/* payment method settings */}
+      <Col xs={24}>
+        <PanelWidth100
+          header={t('memberPlanEdit.paymentConfigs')}
+          bordered
+        >
+          <Row>
+            {/* currency */}
+            <Col xs={12}>
               <Form.ControlLabel>
                 {t('memberPlanEdit.currency')}
               </Form.ControlLabel>
@@ -419,92 +588,7 @@ export function MemberPlanForm({
               />
             </Col>
 
-            {/* minimal monthly amount */}
             <Col xs={12}>
-              <Form.ControlLabel>
-                {t('memberPlanEdit.amountPerMonthMin')}
-              </Form.ControlLabel>
-              <CurrencyInput
-                name="amountPerMonthMin"
-                currency={memberPlan?.currency ?? 'CHF'}
-                centAmount={memberPlan?.amountPerMonthMin || 0}
-                disabled={loading}
-                onChange={centAmount => {
-                  if (!memberPlan) {
-                    return;
-                  }
-                  setMemberPlan({
-                    ...memberPlan,
-                    amountPerMonthMin: centAmount || 0,
-                  });
-                }}
-              />
-              <HelpText>
-                {t('memberplanForm.amountPerMonthMinHelpText')}
-              </HelpText>
-            </Col>
-
-            {/* maximal monthly amount */}
-            <Col xs={12}>
-              <Form.ControlLabel>
-                {t('memberPlanEdit.amountPerMonthMax')}
-              </Form.ControlLabel>
-              <CurrencyInput
-                name="amountPerMonthMax"
-                currency={memberPlan?.currency ?? 'CHF'}
-                centAmount={memberPlan?.amountPerMonthMax ?? null}
-                disabled={loading}
-                onChange={centAmount => {
-                  if (!memberPlan) {
-                    return;
-                  }
-                  setMemberPlan({
-                    ...memberPlan,
-                    amountPerMonthMax: centAmount ?? null,
-                  });
-                }}
-              />
-              <HelpText>
-                {t('memberplanForm.amountPerMonthMaxHelpText')}
-              </HelpText>
-            </Col>
-
-            {/* target monthly amount */}
-            <Col xs={12}>
-              <Form.ControlLabel>
-                {t('memberplanForm.amountPerMonthTarget')}
-              </Form.ControlLabel>
-              <CurrencyInput
-                name="amountPerMonthTarget"
-                currency={memberPlan?.currency ?? 'CHF'}
-                centAmount={memberPlan?.amountPerMonthTarget || 0}
-                disabled={loading}
-                onChange={centAmount => {
-                  if (!memberPlan) {
-                    return;
-                  }
-                  setMemberPlan({
-                    ...memberPlan,
-                    amountPerMonthTarget: centAmount || null,
-                  });
-                }}
-              />
-              <HelpText>
-                {t('memberplanForm.amountPerMonthTargetHelpText')}
-              </HelpText>
-            </Col>
-          </Row>
-        </Panel>
-      </Col>
-
-      {/* payment method settings */}
-      <Col xs={12}>
-        <Panel
-          header={t('memberPlanEdit.paymentConfigs')}
-          bordered
-        >
-          <Row>
-            <Col xs={24}>
               <ListInput
                 value={availablePaymentMethods}
                 disabled={loading}
@@ -603,162 +687,13 @@ export function MemberPlanForm({
                 )}
               </ListInput>
             </Col>
-
-            <Col xs={24}>
-              <Row>
-                <Form.ControlLabel>
-                  {t('memberPlanEdit.successPage')}
-                </Form.ControlLabel>
-                <SelectPage
-                  setSelectedPage={successPageId => {
-                    if (!memberPlan) {
-                      return;
-                    }
-
-                    setMemberPlan({ ...memberPlan, successPageId });
-                  }}
-                  selectedPage={memberPlan?.successPageId}
-                  name="successPageId"
-                />
-              </Row>
-
-              <RowPaddingTop>
-                <Form.ControlLabel>
-                  {t('memberPlanEdit.failPage')}
-                </Form.ControlLabel>
-                <SelectPage
-                  setSelectedPage={failPageId => {
-                    if (!memberPlan) {
-                      return;
-                    }
-
-                    setMemberPlan({ ...memberPlan, failPageId });
-                  }}
-                  selectedPage={memberPlan?.failPageId}
-                  name="failPageId"
-                />
-              </RowPaddingTop>
-
-              <RowPaddingTop>
-                <Form.ControlLabel>
-                  {t('memberplanForm.confirmationPage')}
-                </Form.ControlLabel>
-                <SelectPage
-                  setSelectedPage={confirmationPageId => {
-                    if (!memberPlan) {
-                      return;
-                    }
-
-                    setMemberPlan({ ...memberPlan, confirmationPageId });
-                  }}
-                  selectedPage={memberPlan?.confirmationPageId}
-                  name="failPageId"
-                />
-              </RowPaddingTop>
-              <HelpText>
-                {t('memberplanForm.confirmationPageHelptext')}
-              </HelpText>
-            </Col>
           </Row>
-        </Panel>
+        </PanelWidth100>
       </Col>
 
-      <Col xs={12}>
-        <Panel
-          header={t('memberplanForm.trialSubscription')}
-          bordered
-        >
-          {/* automatically configure trial subscription */}
-          <Row>
-            <Col xs={24}>
-              {isTrialSubscription ?
-                <Alert
-                  icon={<MdCheck />}
-                  severity="success"
-                >
-                  {t('memberplanForm.trialMemberplanAlert')}
-                </Alert>
-              : <Button
-                  startIcon={<MdAutoFixHigh />}
-                  onClick={() =>
-                    setExtendable(
-                      false,
-                      memberPlan ? { ...memberPlan, maxCount: 1 } : undefined
-                    )
-                  }
-                  disabled={isTrialSubscription}
-                  color={'green'}
-                >
-                  {t('memberplanForm.configureTrialBtn')}
-                </Button>
-              }
-            </Col>
-          </Row>
-          <RowPaddingTop>
-            {/* extendable */}
-            <Col xs={12}>
-              <Toggle
-                checked={memberPlan?.extendable}
-                onChange={extendable => setExtendable(extendable)}
-              />
-              <FormControlLabelMarginLeft>
-                {t('memberplanForm.extendableToggle')}
-              </FormControlLabelMarginLeft>
-              <HelpText>{t('memberplanForm.extendableHelpText')}</HelpText>
-            </Col>
-            {/* max count */}
-            <Col xs={12}>
-              <ControlLabel>{maxCountLabel}</ControlLabel>
-              <Input
-                placeholder={maxCountLabel}
-                type={'number'}
-                min={0}
-                value={memberPlan?.maxCount || undefined}
-                onChange={maxCount => {
-                  if (!memberPlan) {
-                    return;
-                  }
-                  setMemberPlan({
-                    ...memberPlan,
-                    maxCount: Number(maxCount) || null,
-                  });
-                }}
-              />
-              <HelpText>{maxCountHelpText}</HelpText>
-            </Col>
-          </RowPaddingTop>
-          <RowPaddingTop>
-            <Col xs={12}>
-              <ControlLabel>{t('memberplanForm.migratePMTitle')}</ControlLabel>
-              <Control
-                name="migrateToTargetPaymentMethodID"
-                block
-                virtualized
-                disabled={loading}
-                data={paymentMethods.map(pm => ({
-                  value: pm.id,
-                  label: pm.name,
-                }))}
-                value={memberPlan?.migrateToTargetPaymentMethodID}
-                accepter={SelectPicker}
-                placement="auto"
-                onChange={migrateToTargetPaymentMethodID =>
-                  setMemberPlan({
-                    ...(memberPlan as FullMemberPlanFragment),
-                    migrateToTargetPaymentMethodID:
-                      migrateToTargetPaymentMethodID || null,
-                  })
-                }
-              />
-              <HelpText>{t('memberplanForm.migratePMHelptext')}</HelpText>
-            </Col>
-          </RowPaddingTop>
-        </Panel>
-      </Col>
-
-      {/* per-periodicity pricing */}
+      {/* pricing */}
       <Col xs={24}>
-        <MemberPlanPeriodicityPricing
+        <MemberPlanPricing
           memberPlan={memberPlan}
           availablePaymentMethods={availablePaymentMethods}
           loading={loading}
